+++ http-api/src/main/java/net/runelite/http/api/item/ItemPrice.java
@@ -31,6 +31,6 @@ public class ItemPrice
 {
 	private int id;
 	private String name;
-	private int price;
-	private int wikiPrice;
+	private long price;
+	private long wikiPrice;
 }
